refactor: consolidate server+agent logging into internal/logging
Output of the monthly consolidation-pass maintenance task. The Sonar
duplication report flagged server/logging/logging.go (87.4% duplicated)
and agent/logging/logging.go (94.5%) as the densest pair in the repo:
the two files were identical except the default InstrumentationName
fallback ("fleet-edr" vs "fleet-edr-agent"), which is dead code since
both call sites pass the name explicitly.

Merged the one true copy into internal/logging (the depguard rule forbids
agent<->server imports; shared code lives in internal/, alongside
internal/observability). Behavior is preserved exactly. The moved test
keeps its two spec: markers so spectrace stays green.

Scrubbed every stale reference to the old paths:
- arch-go.yml: dropped the server/logging platform-allow entry + its rule
  (internal/logging is covered by the existing **.internal.** allowance).
- codecov.yml + sonar-project.properties: removed the agent/logging
  coverage-exclusion (the dir no longer exists; internal/logging is now
  covered by the moved test).

Verified: go build ./..., go test ./internal/logging/..., go vet, arch-go,
the new-code complexity gate, and dash-lint all pass.
