+++ agent/cmd/fleet-edr-agent/main.go
@@ -22,13 +22,13 @@ import (
 	"github.com/fleetdm/edr/agent/enrich"
 	"github.com/fleetdm/edr/agent/enrollment"
 	"github.com/fleetdm/edr/agent/hostid"
-	"github.com/fleetdm/edr/agent/logging"
 	"github.com/fleetdm/edr/agent/metrics"
 	"github.com/fleetdm/edr/agent/proctable"
 	"github.com/fleetdm/edr/agent/queue"
 	"github.com/fleetdm/edr/agent/receiver"
 	"github.com/fleetdm/edr/agent/reconcile"
 	"github.com/fleetdm/edr/agent/uploader"
+	"github.com/fleetdm/edr/internal/logging"
 	"github.com/fleetdm/edr/internal/observability"
 )
