+++ arch-go.yml
@@ -244,7 +244,6 @@ dependenciesRules:
         - "github.com/fleetdm/edr/server/config"
         - "github.com/fleetdm/edr/server/bootstrap"
         - "github.com/fleetdm/edr/server/httpserver"
-        - "github.com/fleetdm/edr/server/logging"
         - "github.com/fleetdm/edr/server/metrics"
         - "github.com/fleetdm/edr/server/attrkeys"
         - "github.com/fleetdm/edr/server/apidocs"
@@ -257,10 +256,6 @@ dependenciesRules:
       internal: *platform_allow
 
   - package: "github.com/fleetdm/edr/server/httpserver"
-    shouldOnlyDependsOn:
-      internal: *platform_allow
-
-  - package: "github.com/fleetdm/edr/server/logging"
     shouldOnlyDependsOn:
       internal: *platform_allow
