+++ docs/maintenance/log.md
@@ -35,4 +35,5 @@ or change the cadence - running tasks with no record is the same as not running 
 2026-06-18  claude-config-audit  done  no committed change (`.claude/` gitignored)  pruned 11 dead allow one-offs from settings.local.json (retired VM `10.0.1.137` + old bearer-token API + Fleet-repo paths); ntfy idle hook valid, no `git push` rule; flagged opsx/openspec skill duplication + broad aws/kubectl breadth for review
 2026-06-18  ai-review-bot-config-audit  done  no findings  `.coderabbit.yaml`: all 7 path_instruction globs match the tree, all 5 disabled-tool CI refs (go/ts/swift-lint, actionlint, osv-scanner) exist, no deprecated keys, multi-platform globs intact; pre-merge trip-rate N/A (CodeRabbit manual-only here)
 2026-06-19  threat-model-and-security-refresh  done  branch `maintenance-threat-model-2026-06-19`  code-driven boundary inventory; threat-model.md: App Control shipped (agent `set_application_control` cmd + detection-content rows), XPC path `extension/`->`shared/XPCEventServer.swift`, named OTLP egress trust assumption; confirmed agent->server is bearer-over-TLS (not mTLS); 5 medium/low gaps (events-at-rest, CSP, mysql-root, harden-runner, CODEOWNERS) unchanged
+2026-06-22  consolidation-pass  done  branch `chore/consolidate-logging-into-internal`  merged 87%/94%-duplicated server+agent `logging.go` into `internal/logging` (scrubbed stale arch-go/codecov/sonar refs); baseline dup 1.3%, cognitive-complexity 5282. Top remaining candidates: `bootstrap.go` cog-26, rbac `"db must not be nil"` x5
