+++ sonar-project.properties
@@ -50,8 +50,6 @@ sonar.javascript.lcov.reportPaths=ui/coverage/lcov.info,test/e2e/coverage/lcov-e
 #                                       file would be 0% by default.
 #   - agent/receiver/**                 CGo bridge; needs a running ESF + NE
 #                                       link, no unit-testable surface.
-#   - agent/logging/**                  thin slog wrapper around stdlib; has
-#                                       no logic worth a test today.
 #   - server/bootstrap/**               process-init wiring, called once.
 #   - server/httpserver/**              net/http server lifecycle wiring.
 #   - test/**                           load-test harness, not production code.
@@ -75,7 +73,7 @@ sonar.javascript.lcov.reportPaths=ui/coverage/lcov.info,test/e2e/coverage/lcov-e
 #                                       runner. No test reaches it, so it
 #                                       always drags the new-code coverage
 #                                       gate to 0% when edited.
-sonar.coverage.exclusions=**/cmd/**/main.go,server/detection/testharness/**,schema/**,extension/edr/**,agent/receiver/**,agent/logging/**,server/bootstrap/**,server/httpserver/**,test/**,tools/**,ui/vite.config.ts
+sonar.coverage.exclusions=**/cmd/**/main.go,server/detection/testharness/**,schema/**,extension/edr/**,agent/receiver/**,server/bootstrap/**,server/httpserver/**,test/**,tools/**,ui/vite.config.ts
 
 # Per-rule suppressions. Sonar's secrets analyser (secrets:S8215) flags the
 # bcrypt password hashes in config/dex/dev-config.yaml + config/dex/demo-config.yaml
